+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.0.8(wip)
+- add type definition for `messagingMessage` event to main stream channel
+
 # 0.0.7
 - Notificationsの型を修正
 - MessagingMessageの型を修正